Remove abstract methods; refactor test
