+++ allensdk/brain_observatory/behavior/session_apis/abcs/behavior_ophys_base.py
@@ -163,48 +163,3 @@ class BehaviorOphysBase(BehaviorBase):
             pupil areas, and frames with likely blinks/outliers.
         """
         raise NotImplementedError()
-
-    @abc.abstractmethod
-    def get_screen_gaze_data(self, include_filtered_data=False) -> Optional[pd.DataFrame]:
-        """
-        Gets screen gaze data
-
-        Parameters
-        ----------
-        include_filtered_data: bool
-            Includes new_* data
-
-        Returns
-        -------
-        pd.DataFrame
-            *_eye_areas: Area of eye (in pixels^2) over time
-            *_pupil_areas: Area of pupil (in pixels^2) over time
-            *_screen_coordinates: y, x screen coordinates (in cm) over time
-            *_screen_coordinates_spherical: y, x screen coordinates (in deg) over time
-            synced_frame_timestamps: synced timestamps for video frames (in sec)
-
-        or None if no eye tracking data
-
-
-        """
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def get_pupil_data(self) -> Optional[pd.DataFrame]:
-        """
-        Gets corneal, eye, and pupil ellipse fit data
-
-        Returns
-        -------
-        pd.DataFrame
-            *_center_x
-            *_center_y
-            *_height
-            *_phi
-            *_width
-
-            where "*" can be "corneal", "pupil" or "eye"
-
-        or None if no eye tracking data
-        """
-        raise NotImplementedError
